pyproject.toml
[tool.poetry]
name = "fhelp"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.104.1"
alembic = "1.12.1"
sqlalchemy = "2.0.23"
psycopg2-binary = "2.9.9"
uvicorn = "0.23.2"
invoke = "2.2.0"
asyncpg = "0.28.0"
pyjwt = "2.8.0"
python-jose = "3.3.0"
redis = "5.0.1"
jinja2 = "^3.1.2"
orjson = "^3.9.10"
pydantic-settings = "^2.0.3"

[tool.poetry.group.dev.dependencies]
black = "^23.10.1"
flake8 = "^6.1.0"
pre-commit = "^3.5.0"
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
httpx = "^0.25.1"
autoflake = "^2.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
